Implement a default bpvec function that works for any param type
Now the default method for bpvec dispatches to bplapply, passing the
param object through to bplapply. So if this is the correct behavior
for a given param class, then there is no need to provide a specific
method just for that param class. The only reason to provide a bpvec
method specific to a given param class is if that param class can
provide a more efficient implementation than simply splitting the
input and dispatching to bplapply.

Hence, now the DoparParam and SnowParam methods of bpvec are now
replaced by the default method, and only MulticoreParam gets its own
bpvec method.