changing (again) how memory use is reported:
There are many ways we could do this. In the interest of consistency
between back-ends I've chosen to go with how BatchJobs reports 
memory use on a worker. BatchJobs calls gc() on exit instead of right
after job evaluation so numbers will (usually) be a littler higher
than what we see for SnowParam or MulticoreParam.


git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/BiocParallel@106700 bc3139a8-67e5-0310-9ffc-ced21a209358